refactor: withdrawal, login info page
