+++ src/pages/mypage/login_info/index.page.tsx
@@ -1,20 +1,27 @@
-'use client';
 import { AUTH_PROVIDER_LABEL, useGetSocialLoginInfo } from '@/apis/member';
+import Header from '@/components/Header/Header';
 import Icon from '@/components/Icon';
 import { css } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
 
+// TODO : 정상화 필요
 function LoginInfoPage() {
   const { data } = useGetSocialLoginInfo();
+
   if (!data) return null;
+
   return (
-    <div className={containerCss}>
-      <Icon name="log-in-information" color="icon.secondary" width={20} height={20} />
-      <div className={textWrapperCss}>
-        <p className={titleCss}>{AUTH_PROVIDER_LABEL[data.provider]}</p>
-        <p className={contentCss}>{data.email}</p>
-      </div>
-    </div>
+    <>
+      <Header rightAction="none" title="로그인 정보" />
+
+      <main className={containerCss}>
+        <Icon name="log-in-information" color="icon.secondary" width={20} height={20} />
+        <div className={textWrapperCss}>
+          <p className={titleCss}>{AUTH_PROVIDER_LABEL[data.provider]}</p>
+          <p className={contentCss}>{data.email}</p>
+        </div>
+      </main>
+    </>
   );
 }
 
@@ -31,6 +38,7 @@ const containerCss = flex({
 });
 
 const textWrapperCss = flex({ flex: 1, flexDirection: 'column', gap: '2px', minWidth: '0' });
+
 const titleCss = css({
   color: 'text.tertiary',
   textStyle: 'body4',
@@ -39,6 +47,7 @@ const titleCss = css({
   whiteSpace: 'nowrap',
   textOverflow: 'ellipsis',
 });
+
 const contentCss = css({
   color: 'text.secondary',
   textStyle: 'body3',
